+++ src/oci/container_engine/models/__init__.py
@@ -13,15 +13,19 @@ from .cluster_summary import ClusterSummary
 from .create_cluster_details import CreateClusterDetails
 from .create_cluster_kubeconfig_content_details import CreateClusterKubeconfigContentDetails
 from .create_node_pool_details import CreateNodePoolDetails
+from .create_node_pool_node_config_details import CreateNodePoolNodeConfigDetails
 from .key_value import KeyValue
 from .kubernetes_network_config import KubernetesNetworkConfig
 from .node import Node
 from .node_error import NodeError
 from .node_pool import NodePool
+from .node_pool_node_config_details import NodePoolNodeConfigDetails
 from .node_pool_options import NodePoolOptions
+from .node_pool_placement_config_details import NodePoolPlacementConfigDetails
 from .node_pool_summary import NodePoolSummary
 from .update_cluster_details import UpdateClusterDetails
 from .update_node_pool_details import UpdateNodePoolDetails
+from .update_node_pool_node_config_details import UpdateNodePoolNodeConfigDetails
 from .work_request import WorkRequest
 from .work_request_error import WorkRequestError
 from .work_request_log_entry import WorkRequestLogEntry
@@ -40,15 +44,19 @@ container_engine_type_mapping = {
     "CreateClusterDetails": CreateClusterDetails,
     "CreateClusterKubeconfigContentDetails": CreateClusterKubeconfigContentDetails,
     "CreateNodePoolDetails": CreateNodePoolDetails,
+    "CreateNodePoolNodeConfigDetails": CreateNodePoolNodeConfigDetails,
     "KeyValue": KeyValue,
     "KubernetesNetworkConfig": KubernetesNetworkConfig,
     "Node": Node,
     "NodeError": NodeError,
     "NodePool": NodePool,
+    "NodePoolNodeConfigDetails": NodePoolNodeConfigDetails,
     "NodePoolOptions": NodePoolOptions,
+    "NodePoolPlacementConfigDetails": NodePoolPlacementConfigDetails,
     "NodePoolSummary": NodePoolSummary,
     "UpdateClusterDetails": UpdateClusterDetails,
     "UpdateNodePoolDetails": UpdateNodePoolDetails,
+    "UpdateNodePoolNodeConfigDetails": UpdateNodePoolNodeConfigDetails,
     "WorkRequest": WorkRequest,
     "WorkRequestError": WorkRequestError,
     "WorkRequestLogEntry": WorkRequestLogEntry,