+++ src/oci/core/models/boot_volume.py
@@ -64,6 +64,10 @@ class BootVolume(object):
             The value to assign to the defined_tags property of this BootVolume.
         :type defined_tags: dict(str, dict(str, object))
 
+        :param system_tags:
+            The value to assign to the system_tags property of this BootVolume.
+        :type system_tags: dict(str, dict(str, object))
+
         :param display_name:
             The value to assign to the display_name property of this BootVolume.
         :type display_name: str
@@ -119,6 +123,7 @@ class BootVolume(object):
             'availability_domain': 'str',
             'compartment_id': 'str',
             'defined_tags': 'dict(str, dict(str, object))',
+            'system_tags': 'dict(str, dict(str, object))',
             'display_name': 'str',
             'freeform_tags': 'dict(str, str)',
             'id': 'str',
@@ -137,6 +142,7 @@ class BootVolume(object):
             'availability_domain': 'availabilityDomain',
             'compartment_id': 'compartmentId',
             'defined_tags': 'definedTags',
+            'system_tags': 'systemTags',
             'display_name': 'displayName',
             'freeform_tags': 'freeformTags',
             'id': 'id',
@@ -154,6 +160,7 @@ class BootVolume(object):
         self._availability_domain = None
         self._compartment_id = None
         self._defined_tags = None
+        self._system_tags = None
         self._display_name = None
         self._freeform_tags = None
         self._id = None
@@ -252,6 +259,32 @@ class BootVolume(object):
         :type: dict(str, dict(str, object))
         """
         self._defined_tags = defined_tags
+
+    @property
+    def system_tags(self):
+        """
+        Gets the system_tags of this BootVolume.
+        System tags for this resource. Each key is predefined and scoped to a namespace.
+        Example: `{\"foo-namespace\": {\"bar-key\": \"value\"}}`
+
+
+        :return: The system_tags of this BootVolume.
+        :rtype: dict(str, dict(str, object))
+        """
+        return self._system_tags
+
+    @system_tags.setter
+    def system_tags(self, system_tags):
+        """
+        Sets the system_tags of this BootVolume.
+        System tags for this resource. Each key is predefined and scoped to a namespace.
+        Example: `{\"foo-namespace\": {\"bar-key\": \"value\"}}`
+
+
+        :param system_tags: The system_tags of this BootVolume.
+        :type: dict(str, dict(str, object))
+        """
+        self._system_tags = system_tags
 
     @property
     def display_name(self):