+++ src/oci/database/models/autonomous_data_warehouse_summary.py
@@ -545,8 +545,6 @@ class AutonomousDataWarehouseSummary(object):
         Defined tags for this resource. Each key is predefined and scoped to a namespace.
         For more information, see `Resource Tags`__.
 
-        Example: `{\"Operations\": {\"CostCenter\": \"42\"}}`
-
         __ https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm
 
 
@@ -561,8 +559,6 @@ class AutonomousDataWarehouseSummary(object):
         Sets the defined_tags of this AutonomousDataWarehouseSummary.
         Defined tags for this resource. Each key is predefined and scoped to a namespace.
         For more information, see `Resource Tags`__.
-
-        Example: `{\"Operations\": {\"CostCenter\": \"42\"}}`
 
         __ https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm