+++ src/oci/load_balancer/models/backend.py
@@ -238,6 +238,8 @@ class Backend(object):
         Whether the load balancer should treat this server as a backup unit. If `true`, the load balancer forwards no ingress
         traffic to this backend server unless all other backend servers not marked as \"backup\" fail the health check policy.
 
+        **Note:** You cannot add a backend server marked as `backup` to a backend set that uses the IP Hash policy.
+
         Example: `false`
 
 
@@ -252,6 +254,8 @@ class Backend(object):
         Sets the backup of this Backend.
         Whether the load balancer should treat this server as a backup unit. If `true`, the load balancer forwards no ingress
         traffic to this backend server unless all other backend servers not marked as \"backup\" fail the health check policy.
+
+        **Note:** You cannot add a backend server marked as `backup` to a backend set that uses the IP Hash policy.
 
         Example: `false`